Refactor tasks to avoid importing unneeded things when not running e2e tasks
