+++ tasks/e2e_framework/aws/ecs.py
@@ -1,9 +1,8 @@
 from invoke.context import Context
 from invoke.exceptions import Exit
 from invoke.tasks import task
-from pydantic import ValidationError
 
-from tasks.e2e_framework import config, doc, tool
+from tasks.e2e_framework import doc, tool
 from tasks.e2e_framework.aws import doc as aws_doc
 from tasks.e2e_framework.aws.common import get_aws_wrapper
 from tasks.e2e_framework.aws.deploy import deploy
@@ -77,6 +76,9 @@ def create_ecs(
 
 def _show_connection_message(ctx: Context, config_path: str | None, full_stack_name: str):
     import pyperclip
+    from pydantic import ValidationError
+
+    from tasks.e2e_framework import config
 
     outputs = tool.get_stack_json_outputs(ctx, full_stack_name)
     cluster_name = outputs["dd-Cluster-ecs"]["clusterName"]
