+++ tasks/e2e_framework/gcp/vm.py
@@ -1,10 +1,8 @@
 from invoke.context import Context
 from invoke.exceptions import Exit
 from invoke.tasks import task
-from pydantic_core._pydantic_core import ValidationError
 
-from tasks.e2e_framework import config, doc, tool
-from tasks.e2e_framework.config import get_full_profile_path
+from tasks.e2e_framework import doc, tool
 from tasks.e2e_framework.deploy import deploy
 from tasks.e2e_framework.destroy import destroy
 from tasks.e2e_framework.gcp import doc as gcp_doc
@@ -75,10 +73,14 @@ def create_vm(
     Create a new virtual machine on gcp.
     """
 
+    from pydantic_core._pydantic_core import ValidationError
+
+    from tasks.e2e_framework import config
+
     try:
         cfg = config.get_local_config(config_path)
     except ValidationError as e:
-        raise Exit(f"Error in config {get_full_profile_path(config_path)}") from e
+        raise Exit(f"Error in config {config.get_full_profile_path(config_path)}") from e
 
     if not cfg.get_gcp().publicKeyPath:
         raise Exit("The field `gcp.publicKeyPath` is required in the config file")
