+++ tasks/e2e_framework/localpodman/vm.py
@@ -1,9 +1,8 @@
 from invoke.context import Context
 from invoke.exceptions import Exit
 from invoke.tasks import task
-from pydantic_core._pydantic_core import ValidationError
 
-from tasks.e2e_framework import config, doc
+from tasks.e2e_framework import doc
 from tasks.e2e_framework.deploy import deploy
 from tasks.e2e_framework.destroy import destroy
 from tasks.e2e_framework.tool import add_known_host as add_known_host_func
@@ -46,6 +45,9 @@ def create_vm(
     """
     Create a new virtual machine on local podman.
     """
+    from pydantic_core._pydantic_core import ValidationError
+
+    from tasks.e2e_framework import config
 
     try:
         cfg = config.get_local_config(config_path)
