+++ tasks/e2e_framework/setup/agent.py
@@ -1,0 +1,41 @@
+from tasks.e2e_framework.config import Config
+from tasks.e2e_framework.tool import ask, warn
+
+
+def setup_agent_config(config):
+    if config.configParams.agent is None:
+        config.configParams.agent = Config.Params.Agent(
+            apiKey=None,
+            appKey=None,
+        )
+    # API key
+    if config.configParams.agent.apiKey is None:
+        config.configParams.agent.apiKey = "0" * 32
+    default_api_key = config.configParams.agent.apiKey
+    while True:
+        config.configParams.agent.apiKey = default_api_key
+        apiKey = ask(f"🐶 Datadog API key - default [{_get_safe_dd_key(config.configParams.agent.apiKey)}]: ")
+        if len(apiKey) > 0:
+            config.configParams.agent.apiKey = apiKey
+        if len(config.configParams.agent.apiKey) == 32:
+            break
+        warn(f"Expecting API key of length 32, got {len(config.configParams.agent.apiKey)}")
+    # APP key
+    if config.configParams.agent.appKey is None:
+        config.configParams.agent.appKey = "0" * 40
+    default_app_key = config.configParams.agent.appKey
+    while True:
+        config.configParams.agent.appKey = default_app_key
+
+        app_Key = ask(f"🐶 Datadog APP key - default [{_get_safe_dd_key(config.configParams.agent.appKey)}]: ")
+        if len(app_Key) > 0:
+            config.configParams.agent.appKey = app_Key
+        if len(config.configParams.agent.appKey) == 40:
+            break
+        warn(f"Expecting APP key of length 40, got {len(config.configParams.agent.appKey)}")
+
+
+def _get_safe_dd_key(key: str) -> str:
+    if key == "0" * len(key):
+        return key
+    return "*" * len(key)
